refactor: Rename and reorder pie chart items
