+++ lib/app/home/widgets/dashboard_cards.dart
@@ -3,7 +3,7 @@ import 'package:monekin/app/stats/stats_page.dart';
 import 'package:monekin/app/stats/widgets/balance_bar_chart.dart';
 import 'package:monekin/app/stats/widgets/finance_health/finance_health_main_info.dart';
 import 'package:monekin/app/stats/widgets/fund_evolution_line_chart.dart';
-import 'package:monekin/app/stats/widgets/movements_distribution/chart_by_categories.dart';
+import 'package:monekin/app/stats/widgets/movements_distribution/pie_chart_by_categories.dart';
 import 'package:monekin/core/models/date-utils/date_period_state.dart';
 import 'package:monekin/core/presentation/responsive/breakpoints.dart';
 import 'package:monekin/core/presentation/responsive/responsive_row_column.dart';
@@ -69,7 +69,7 @@ class DashboardCards extends StatelessWidget {
               const SizedBox(height: 16),
               CardWithHeader(
                 title: t.stats.by_categories,
-                body: ChartByCategories(datePeriodState: dateRangeService),
+                body: PieChartByCategories(datePeriodState: dateRangeService),
                 footer: CardFooterWithSingleButton(
                   onButtonClick: () => RouteUtils.pushRoute(
                     context,
