+++ lib/app/stats/stats_page.dart
@@ -3,7 +3,7 @@ import 'package:monekin/app/stats/widgets/balance_bar_chart.dart';
 import 'package:monekin/app/stats/widgets/finance_health_details.dart';
 import 'package:monekin/app/stats/widgets/fund_evolution_line_chart.dart';
 import 'package:monekin/app/stats/widgets/income_expense_comparason.dart';
-import 'package:monekin/app/stats/widgets/movements_distribution/chart_by_categories.dart';
+import 'package:monekin/app/stats/widgets/movements_distribution/pie_chart_by_categories.dart';
 import 'package:monekin/app/stats/widgets/movements_distribution/tags_stats.dart';
 import 'package:monekin/core/database/services/account/account_service.dart';
 import 'package:monekin/core/models/date-utils/date_period_state.dart';
@@ -169,7 +169,7 @@ class _StatsPageState extends State<StatsPage> {
                 buildContainerWithPadding([
                   CardWithHeader(
                     title: t.stats.by_categories,
-                    body: ChartByCategories(
+                    body: PieChartByCategories(
                       datePeriodState: dateRangeService,
                       showList: true,
                       initialSelectedType: TransactionType.E,
